backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.0
pydantic==2.4.2
python-multipart==0.0.6
numpy==1.21.2
torch==1.9.0
transformers==4.11.3
whisper>=1.0.0
sqlalchemy>=2.0.0
python-jose>=3.3.0
passlib>=1.7.4
aiohttp>=3.8.5
asyncio>=3.4.3
pandas>=2.0.0
scikit-learn>=1.3.0 